LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.8_Literals/7.8.5_Regular_Expression_Literals/S7.8.5_A1.2_T1-expected.txt

CONSOLE MESSAGE: line 84: SyntaxError: Parse error
S7.8.5_A1.2_T1

PASS Expected parsing failure

TEST COMPLETE
